+++ HISTORY.md
@@ -3,8 +3,9 @@
 ## in development:
 * **New Features:**
     * new primitive in "looks": NEW COSTUME from a list of pixels and dimensions, allowing CURRENT
-    * new primitige in "sound": NEW SOUND from a list of samples
+    * new primitive in "sound": NEW SOUND from a list of samples
     * added selectors for sprites' and the stage's bounding box (LEFT, RIGHT, TOP, BOTTOM) to MY dropdown
+    * new experimental entry for "green flag pressed" in the BROADCAST block's dropdown when the shift key is pressed
 * **Notable Changes:**
     * running STOP ALL now also toggles (pauses and resumes) all generic WHEN hat blocks (just like pressing the stop button)
     * changed default name for new costumes created with STRETCH etc. to localized 'costume'
@@ -13,6 +14,9 @@
     * assert that dimensions given for STRETCH are finite numbers (avoid crash)
 * **Translation Updates:**
     * German
+
+### 2019-10-21
+* blocks, threads: new experimental entry for "green flag pressed" in the BROADCAST block's dropdown when the shift key is pressed
 
 ### 2019-10-20
 * objects, threads: added "new sound" from list of samples primitive reporter to "sound" category